requirements.txt
Django==4.2
djangorestframework==3.15.2
gunicorn==23.0.0
drf-spectacular==0.28.0
Pillow==11.1.0
